+++ packages/graphback-codegen-resolvers/src/formatters/apollo/resolverFormatter.ts
@@ -1,5 +1,5 @@
 import { ResolverGeneratorPluginOptions } from '../../plugin/ResolverGeneratorPlugin';
-import { createCustomResolversIndex as createCustomResolversIndexJS, createResolversIndexJS as createResolversIndexJS, resolverFileTemplateJS as resolverJSFileTemplate, rootResolversIndexJS } from './jsResolverFormatter';
+import { createCustomResolversIndexJS, createResolversIndexJS as createResolversIndexJS, resolverFileTemplateJS as resolverJSFileTemplate, rootResolversIndexJS } from './jsResolverFormatter';
 import { createResolversIndexTS, resolverFileTemplateTS as resolverTSFileTemplate, rootResolversIndexTS } from './tsResolverFormatter';
 
 const mapResolverKeyValueTemplates = (resolvers: any) => {
